fix(core): widen ACK position wire field to 64 bits
ACK frames carried the cumulative byte position as a 32-bit unsigned integer, then decoded it through (int) in WriteChannel.OnAck. Once a single channel had received more than 2 GiB cumulative frame bytes the cast went negative, _ackedPos froze, and the writer permanently blocked on _spaceAvailable — a silent deadlock on long-running, high-throughput channels (e.g. ~21 s at 100 MB/s).

Widen the wire payload from 4 to 8 bytes (uint64 big-endian), change IChannelOwner.SendAck to ulong, WriteChannel.OnAck to long, and adjust ReadChannel encode, StreamMultiplexer routing/init-ack, and WriteAckFrame helper to match. Update the framing-protocol doc and add a round-trip regression test for positions above uint.MaxValue.

+++ src/NetConduit/Internal/IChannelOwner.cs
@@ -21,5 +21,5 @@ internal interface IChannelOwner
     /// Used by ReadChannel to inform the remote WriteChannel how far the consumer
     /// has consumed, so PrepareReplay on reconnect skips already-delivered bytes.
     /// </summary>
-    void SendAck(ushort channelIndex, uint consumedPosition);
+    void SendAck(ushort channelIndex, ulong consumedPosition);
 }
